requirements.txt
ccxt>=4.2
websockets>=12.0
aiohttp>=3.9
pydantic>=2.7
python-dotenv>=1.0
prometheus-client>=0.20